[parallel] force the gathered information to be of type int
This fixes a compilation error encountered with YaspGrid and the Dune
master. Since the return type of mappers has been generalized, an
implicit conversion to int does not happen anymore.

Reviewed by gruenich.


git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/trunk@13479 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ dumux/linear/vertexborderlistfromgrid.hh
@@ -81,7 +81,7 @@ public:
     void gather(MessageBufferImp &buff, const EntityType &e) const
     {
         buff.write(gridView_.comm().rank());
-        buff.write(map_.map(e));
+        buff.write(static_cast<int>(map_.map(e)));
     }
 
     template<class MessageBufferImp, class EntityType>
